Add a sample of how to import an EDK2 sample
So far, VisualEfi came with its own samples. But EDK2 has its own.
FtdiUsbSerialDxe is an example project file that imports an existing
EDK2 sample. VsGlue.c is needed for such projects to redirect execution
to the correct entrypoint name (EDK2's build environment does this on
its own with special perl/python scripts editing the template source
files. A simple 2-function file is not the end of the world).